add flash messages and clean up css

+++ app/controllers/climbers_controller.rb
@@ -34,7 +34,8 @@ class ClimbersController < ApplicationController
     # same code from new action, for people trying to hack in a post request
     @climber = Climber.new(climb_params)
     if @climber.save
-      pending.first.destroy
+      pending.first.destroy unless pending.empty?
+      flash[:notice] = 'Account successfully created!'
       redirect_to(@climber)
     else
       @email = pending.first.email unless pending.empty?
